Rework of Plot revert so adjusting it's timer will not affect plot claiming (snapshots).

+++ src/com/palmergames/bukkit/towny/tasks/RepeatingTimerTask.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 
 import com.palmergames.bukkit.towny.NotRegisteredException;
 import com.palmergames.bukkit.towny.TownyLogger;
+import com.palmergames.bukkit.towny.TownySettings;
 import com.palmergames.bukkit.towny.object.PlotBlockData;
 import com.palmergames.bukkit.towny.object.TownBlock;
 import com.palmergames.bukkit.towny.object.TownyRegenAPI;
@@ -15,18 +16,29 @@ public class RepeatingTimerTask extends TownyTimerTask {
 		super(universe);
 	}
 	
+	private Long timerCounter = 0L;
+	
 	@Override
 	public void run() {
+		timerCounter++;
+		
 		// Perform a single block regen in each regen area, if any are left to do.
 		if (TownyRegenAPI.hasPlotChunks()) {
-			for (PlotBlockData plotChunk : new ArrayList<PlotBlockData>(TownyRegenAPI.getPlotChunks().values())) {
-				if (!plotChunk.restoreNextBlock()) {
-					TownyRegenAPI.deletePlotChunk(plotChunk);	
-					TownyRegenAPI.deletePlotChunkSnapshot(plotChunk);
+			// only execute if the correct amount of time has passed.
+			if (Math.max(1L, TownySettings.getPlotManagementSpeed()) == timerCounter) {
+				for (PlotBlockData plotChunk : new ArrayList<PlotBlockData>(TownyRegenAPI.getPlotChunks().values())) {
+					if (!plotChunk.restoreNextBlock()) {
+						TownyRegenAPI.deletePlotChunk(plotChunk);	
+						TownyRegenAPI.deletePlotChunkSnapshot(plotChunk);
+					}
 				}
+				timerCounter = 0L;
 			}
 		}
 		
+		/**
+		 * The following actions should be performed every second.
+		 */
 		// Take a snapshot of the next townBlock and save.
 		if (TownyRegenAPI.hasWorldCoords()) {
 			try {
